+++ src/main/java/dev/root101/clean/core/repo/DefaultCRUDRepo.java
@@ -138,10 +138,10 @@ public class DefaultCRUDRepo<Domain extends DomainObject<ID>, Entity, ID, Extern
     }
 
     @Override
-    public int count() throws RuntimeException {
+    public long count() throws RuntimeException {
         firePropertyChange(BEFORE_COUNT, null, null);
 
-        int c = externalRepo.count();
+        long c = externalRepo.count();
 
         firePropertyChange(AFTER_COUNT, null, c);
